branches/jAER2: update Events according to spec. Implement RawEvent decoding for main event types (Block Event will follow).
git-svn-id: svn://svn.code.sf.net/p/jaer/code@4435 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ branches/jAER_2.0/src/main/java/net/sf/jaer2/eventio/events/SampleEvent.java
@@ -3,18 +3,18 @@ package net.sf.jaer2.eventio.events;
 public class SampleEvent extends Event {
 	private static final long serialVersionUID = 1922113968889799454L;
 
-	private short type;
+	private byte type;
 	private int sample;
 
 	public SampleEvent(final int ts) {
 		super(ts);
 	}
 
-	public short getType() {
+	public byte getType() {
 		return type;
 	}
 
-	public void setType(final short type) {
+	public void setType(final byte type) {
 		this.type = type;
 	}
